test SimpleChart
