--- src/__tests__/ExampleUnit.test.ts
@@ -1,5 +1,0 @@
-describe('example unit test', () => {
-    test('always passes', () => {
-        expect(true).toBe(true);
-    })
-})
