+++ src/test/test-setup.ts
@@ -1,0 +1,4 @@
+import { cleanup } from '@testing-library/react'
+import { beforeEach } from 'vitest/dist/index.js'
+
+beforeEach(cleanup);
