+++ vite.config.ts
@@ -4,9 +4,11 @@ import react from '@vitejs/plugin-react'
 // https://vitejs.dev/config/
 export default defineConfig({
   test: {
+    setupFiles: './src/test/test-setup.ts',
     coverage: {
-      provider: 'v8' // or 'v8'
+      provider: 'v8' // or istanbul
     },
+    environment: 'jsdom',
   },
   plugins: [react()],
 })
